Split hook and wrappers
